+++ Mods ChoGGi/Faster UI/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2744055836",
 	"pops_any_uuid", "7cb99512-05c3-44ba-b4b8-1fcb37540cab",
 	"lua_revision", 1007000, -- Picard
-	"version", 3,
+	"version", 4,
 	"version_major", 0,
-	"version_minor", 3,
+	"version_minor", 4,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -16,10 +16,10 @@ return PlaceObj("ModDef", {
 	"TagInterface", true,
 	"description", [[
 Remove some animations: build menu/pins/overview/map switch.
-
+See mod options to enable/disable each speed up.
 
 Known Issues:
-Still a slight delay when closing build menu and opening pins :(
-It might be a bit jarring when going between the map overview (I can make it a mod option).
+Still a very slight delay when closing build menu and opening pins :(
+It might be a bit jarring when going between the map overview.
 ]],
 })