+++ Mods ChoGGi/Fix Borked Mod Building Template/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2552439288",
 	"pops_any_uuid", "8ac9c513-fbe0-4c94-88ab-413c1c0ff485",
 	"lua_revision", 1007000, -- Picard
-	"version", 3,
+	"version", 4,
 	"version_major", 0,
-	"version_minor", 3,
+	"version_minor", 4,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {