+++ Mods ChoGGi/Standing Unlocks Sponsor Buildings/Code/Script.lua
@@ -1,5 +1,10 @@
 -- See LICENSE for terms
 
+if not g_AvailableDlc.gagarin then
+	print(CurrentModDef.title, ": Space Race DLC not installed!")
+	return
+end
+
 local mod_ShowNotification
 local mod_MinStanding